requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-APScheduler==1.13.1
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
pywhatkit==5.4
twilio==8.11.0
python-dotenv==1.0.0
cryptography==41.0.7
python-magic==0.4.27
Jinja2==3.1.2
email-validator==2.1.0
WTForms==3.1.1
google-search-results==2.4.2
gunicorn>=22.0
psycopg2-binary>=2.9
